Got delete working
